test(core): make the CUDA arch canary run under cargo nextest
build.rs's own #[cfg(test)] mod tests -- including the sm_75-floor
regression canary for #255/#196 -- was never collected by any test
runner: a build script is not a test target. Move the pure
cuda_gpu_targets_from_raw/normalize_cuda_gpu_targets parsing plus that
canary into src/cuda_targets.rs, include!'d verbatim by build.rs (it
cannot depend on the crate it configures) and compiled into the crate
as an ordinary #[cfg(test)] module, so the canary now runs as part of
openasr-core's cargo nextest binary.
