+++ crates/openasr-core/build.rs
@@ -1513,47 +1513,20 @@ fn hip_gpu_targets() -> String {
         })
 }
 
+// Pure CUDA-arch-list parsing/defaults (`cuda_gpu_targets_from_raw`,
+// `normalize_cuda_gpu_targets`, `DEFAULT_CUDA_GPU_TARGETS`) live in
+// `src/cuda_targets.rs`, not here: that file is `include!`d verbatim (this
+// build script cannot depend on the crate it configures) and is ALSO
+// compiled as an ordinary `#[cfg(test)] mod` of this crate (see `lib.rs`),
+// so its regression canary -- guarding the class of bug fixed in #255/#196
+// where OpenASR's own CUDA arch-list default silently narrowed below
+// vendored ggml's sm_75 floor -- actually runs under `cargo nextest`. A
+// `#[cfg(test)] mod tests` living only in a build script is never collected
+// by any test runner, so before this the canary had no teeth.
+include!("src/cuda_targets.rs");
+
 fn cuda_gpu_targets() -> String {
     cuda_gpu_targets_from_raw(env::var("OPENASR_CUDA_GPU_TARGETS").ok().as_deref())
-}
-
-fn cuda_gpu_targets_from_raw(raw: Option<&str>) -> String {
-    raw.map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(normalize_cuda_gpu_targets)
-        .unwrap_or_else(|| {
-            // Default CUDA arch list for a released binary: sm_75 through sm_90.
-            // Floor is 75 (Turing: RTX 20xx, GTX 16xx, T4, 2080 Ti), not 70
-            // (Volta) or lower, because CUDA 13 removed device-code generation
-            // for Volta/Pascal/Maxwell outright -- those need a CUDA 12
-            // toolchain, which is a separate build leg (tracked, not this one).
-            // 75 itself is deprecated-but-still-buildable on CUDA 13.2: nvcc
-            // warns, but still emits working sm_75 SASS and PTX. No explicit
-            // `-real`/`-virtual` suffix means CMake emits both device code and
-            // PTX for every listed number, so this default also carries
-            // 75-virtual PTX that the CUDA driver JITs forward to any newer,
-            // unlisted architecture at load time -- the same forward-compat
-            // mechanism vendored ggml's own (non-native) CMakeLists.txt default
-            // relies on. Upper bound stops at 90 (Hopper): 120 (Blackwell)
-            // needs CUDA 12.8+ and a CMake new enough for the `f`/`a` suffix
-            // regex, which this default does not assume. Override with
-            // OPENASR_CUDA_GPU_TARGETS for a narrower/wider/newer set.
-            "75;80;86;89;90".to_string()
-        })
-}
-
-fn normalize_cuda_gpu_targets(raw: &str) -> String {
-    raw.split([',', ';', ' '])
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(|value| {
-            value
-                .strip_prefix("sm_")
-                .or_else(|| value.strip_prefix("SM_"))
-                .unwrap_or(value)
-        })
-        .collect::<Vec<_>>()
-        .join(";")
 }
 
 fn prepare_windows_hip_sdk_shim(target: &str, hip_path: &Path, out_dir: &Path) -> PathBuf {
@@ -1805,9 +1778,8 @@ mod tests {
     use std::collections::HashMap;
 
     use super::{
-        CudaTuning, HipTuning, cuda_gpu_targets_from_raw, ios_deployment_target_from,
-        macos_deployment_target_from, parse_bool_env, resolve_ggml_native_enabled, target_is_x86,
-        version_at_least,
+        CudaTuning, HipTuning, ios_deployment_target_from, macos_deployment_target_from,
+        parse_bool_env, resolve_ggml_native_enabled, target_is_x86, version_at_least,
     };
 
     #[test]
@@ -1884,49 +1856,11 @@ mod tests {
         assert!(!target_is_x86("aarch64-apple-darwin"));
     }
 
-    #[test]
-    fn cuda_targets_default_to_common_cloud_and_consumer_arches() {
-        assert_eq!(cuda_gpu_targets_from_raw(None), "75;80;86;89;90");
-        assert_eq!(cuda_gpu_targets_from_raw(Some("   ")), "75;80;86;89;90");
-    }
-
-    #[test]
-    fn cuda_targets_accept_common_arch_spellings() {
-        assert_eq!(
-            cuda_gpu_targets_from_raw(Some("sm_80,86; SM_89 90")),
-            "80;86;89;90"
-        );
-    }
-
-    #[test]
-    fn cuda_default_targets_are_not_narrower_than_vendored_ggml_default() {
-        // Regression guard for the class of bug in #255/#196: OpenASR's own
-        // default CUDA arch list silently excluded sm_75 (Turing), so a
-        // released CUDA binary hit "no kernel image available" on every
-        // 2080 Ti / T4 / RTX 20xx / GTX 16xx card. Cross-check against
-        // vendored ggml's own (non-native) CMakeLists.txt default, which is
-        // the actual upstream source of truth this default is trying to
-        // track: if a future ggml bump drops or narrows its own sm_75 floor,
-        // this test's first assertion fails and flags that our comment/
-        // reasoning in `cuda_gpu_targets_from_raw` needs re-deriving; if
-        // OpenASR's own default regresses below that floor, the second
-        // assertion fails instead.
-        let vendored_ggml_cmake =
-            include_str!("third_party/openasr-ggml/src/ggml-cuda/CMakeLists.txt");
-        assert!(
-            vendored_ggml_cmake.contains("75-virtual"),
-            "vendored ggml CMakeLists.txt no longer defaults to sm_75; re-check \
-             whether OpenASR's own CUDA_GPU_TARGETS floor of 75 is still \
-             upstream-aligned before changing it"
-        );
-        assert!(
-            cuda_gpu_targets_from_raw(None)
-                .split(';')
-                .any(|arch| arch == "75"),
-            "OpenASR's default CUDA arch list narrowed below upstream ggml's \
-             own sm_75 floor"
-        );
-    }
+    // The CUDA-arch-target parsing/default tests (including the sm_75-floor
+    // regression canary for #255/#196) moved to `src/cuda_targets.rs`, which
+    // this build script `include!`s -- see that file's doc comment for why:
+    // a `#[cfg(test)] mod` defined only in a build script is never collected
+    // by `cargo test`/`cargo nextest`, so it ran nowhere.
 
     #[test]
     fn cuda_tuning_defaults_match_fast_compile_safe_defaults() {
